rename helper function file (QA)
