+++ client/src/components/QAhelperFunctions.js
